refactor: extract unit to components directory
